Consolidating code; removing helpers file
