+++ src/services/behaviors/index.ts
@@ -7,8 +7,17 @@
  */
 
 import * as Dom from 'hyperview/src/services/dom';
-import type { HvComponentOnUpdate, UpdateAction } from 'hyperview/src/types';
-import { ACTIONS } from 'hyperview/src/types';
+import {
+  ACTIONS,
+  BEHAVIOR_ATTRIBUTES,
+  NAV_ACTIONS,
+  UPDATE_ACTIONS,
+} from 'hyperview/src/types';
+import type {
+  HvComponentOnUpdate,
+  NavAction,
+  UpdateAction,
+} from 'hyperview/src/types';
 import { shallowCloneToRoot } from 'hyperview/src/services';
 
 /**
@@ -109,6 +118,9 @@ export const performUpdate = (
   return shallowCloneToRoot(targetElement);
 };
 
+/**
+ * Trigger all behaviors matching the given name
+ */
 export const trigger = (
   name: string,
   element: Element,
@@ -139,4 +151,72 @@ export const trigger = (
   });
 };
 
-export { createActionHandler, triggerBehaviors } from './helpers';
+/**
+ * Trigger a set of pre-filtered behaviors
+ */
+export const triggerBehaviors = (
+  element: Element,
+  behaviors: Element[],
+  onUpdate: HvComponentOnUpdate,
+) => {
+  behaviors.forEach(behaviorElement => {
+    const handler = createActionHandler(behaviorElement, onUpdate);
+    if (
+      behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.IMMEDIATE) === 'true'
+    ) {
+      handler(element);
+    } else {
+      setTimeout(() => handler(element), 0);
+    }
+  });
+};
+
+export const createActionHandler = (
+  behaviorElement: Element,
+  onUpdate: HvComponentOnUpdate,
+) => {
+  const action =
+    behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.ACTION) ||
+    NAV_ACTIONS.PUSH;
+  if (Object.values(NAV_ACTIONS).indexOf(action as NavAction) >= 0) {
+    return (element: Element) => {
+      const href = behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.HREF);
+      const targetId = behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.TARGET);
+      const showIndicatorId = behaviorElement.getAttribute(
+        BEHAVIOR_ATTRIBUTES.SHOW_DURING_LOAD,
+      );
+      const delay = behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.DELAY);
+      onUpdate(href, action, element, { delay, showIndicatorId, targetId });
+    };
+  }
+  if (
+    action === ACTIONS.RELOAD ||
+    Object.values(UPDATE_ACTIONS).indexOf(action as UpdateAction) >= 0
+  ) {
+    return (element: Element) => {
+      const href = behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.HREF);
+      const verb = behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.VERB);
+      const targetId = behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.TARGET);
+      const showIndicatorIds = behaviorElement.getAttribute(
+        BEHAVIOR_ATTRIBUTES.SHOW_DURING_LOAD,
+      );
+      const hideIndicatorIds = behaviorElement.getAttribute(
+        BEHAVIOR_ATTRIBUTES.HIDE_DURING_LOAD,
+      );
+      const delay = behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.DELAY);
+      const once = behaviorElement.getAttribute(BEHAVIOR_ATTRIBUTES.ONCE);
+      onUpdate(href, action, element, {
+        behaviorElement,
+        delay,
+        hideIndicatorIds,
+        once,
+        showIndicatorIds,
+        targetId,
+        verb,
+      });
+    };
+  }
+  // Custom behavior
+  return (element: Element) =>
+    onUpdate(null, action, element, { behaviorElement, custom: true });
+};
